requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.8
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-formtools==2.2
django-guardian==2.3.0
django-import-export==2.3.0
django-redis==4.12.1
django-reversion==3.0.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.0
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.2.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.8
pycryptodomex==3.9.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-alipay-sdk==2.1.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.9
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
